test: share the per-run state isolation across every host-booting suite
Issue #259. The Execution API was not the only suite persisting into its build
output. Measured on this machine: Presentation.AgentHub.Tests had accumulated a
40 KB conversation database and Presentation.Common.Tests 32 KB, by the same
route that reached 540 KB in the Execution API suite and eventually made it
return bad requests that read as a product regression.

Three sites is where copying stops being reasonable, so the mechanism moves to
a new Tests.Common support library and the Execution API's copy becomes a call
into it. That project is deliberately marked IsTestProject=false: it holds no
tests and must not be handed to the runner.

Each assembly keeps its own one-line [ModuleInitializer]. The initializer
cannot live in the shared library, because a module initializer runs when ITS
module loads and the library's module may not load until after the consuming
assembly has already booted a host. The consuming assembly's own initializer is
the only hook guaranteed to run first.

Checked and NOT affected: Presentation.EvalRunner.Tests,
Presentation.FoundryHost.Tests and Presentation.BundleApi.Tests write no
database to their build output.

Verified: both newly covered suites pass, and two consecutive full runs of each
leave zero files in their build-output data directory.

Unrelated pre-existing flakiness observed while verifying, filed separately:
AgentHub's telemetry metric-emission tests fail intermittently when that suite
runs on its own — three failed on unmodified main under the same command, and
the failing set differs run to run.

Closes #259
