+++ src/Content/Tests/Presentation.ExecutionApi.Tests/Presentation.ExecutionApi.Tests.csproj
@@ -33,6 +33,7 @@
     <ProjectReference Include="..\..\Presentation\Presentation.ExecutionApi\Presentation.ExecutionApi.csproj" />
     <ProjectReference Include="..\..\Application\Application.AI.Common\Application.AI.Common.csproj" />
     <ProjectReference Include="..\..\Domain\Domain.AI\Domain.AI.csproj" />
+    <ProjectReference Include="..\Tests.Common\Tests.Common.csproj" />
   </ItemGroup>
 
 </Project>
