+++ src/Content/Tests/Tests.Common/Tests.Common.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <IsPackable>false</IsPackable>
+    <!-- Not a test project: it contains no tests and must not be handed to the runner.
+         It is support code shared BY test projects. -->
+    <IsTestProject>false</IsTestProject>
+    <RootNamespace>Tests.Common</RootNamespace>
+    <Configurations>Debug;Release;Debug-AgentHub-All;Debug-AgentHub-WebUI;Debug-AgentHub-Dashboard</Configurations>
+  </PropertyGroup>
+
+</Project>
